requirements.txt
aiogram~=3.21.0
python-dotenv~=1.1.1
SQLAlchemy~=2.0.43